+++ src/client/render/PlayerRenderer.cpp
@@ -50,7 +50,7 @@ namespace render {
 			this->sprite_transforms.push_back(sf::Transform(sprite_transforms[0]).translate(20.f,20.f));				//Transform
 
 			//Enter Cards (position 2 to 16 in sprites)
-			for (i=0;i<14 and i<tobuild.size();i++){
+			for (i=0;i<14 and i<(int) tobuild.size();i++){
 				cRenderer = new DevelopmentCardRenderer(tobuild[i],sf::Transform(sprite_transforms[0]).translate(350.f+70.f*(i%7),48.f+127.f*(i/7)).scale(0.2f,0.2f).scale(1.f,(431.f/375.f)));
 				this->devCardRenderers.push_back(cRenderer);	//Card Renderer
 			}
@@ -105,7 +105,7 @@ namespace render {
 			this->sprite_transforms.push_back(sf::Transform(sprite_transforms[0]).translate(10.f,50.f));				//Transform
 
 			//Enter Cards (position 2 to 9 in sprites)
-			for (i=0;i<7 and i<drafted.size();i++){
+			for (i=0;i<7 and i<(int) drafted.size();i++){
 				cRenderer = new DevelopmentCardRenderer(drafted[i],sf::Transform(sprite_transforms[0]).translate(350.f+180.f*(i),10.f).scale(0.35f,0.35f).scale(1.f,(431.f/375.f)));
 				this->devCardRenderers.push_back(cRenderer);	//Card Renderer
 			}
@@ -139,11 +139,11 @@ namespace render {
 			this->sprite_transforms.push_back(sf::Transform(sprite_transforms[0]).translate(50.f,50.f).scale(2.f,2.f).scale(1.f,(431.f/375.f)));				//Transform
 
 			//Enter Cards (position 2 to 58 in sprites)
-			for (i=0;i<28 and i<built.size();i++){
+			for (i=0;i<28 and i<(int) built.size();i++){
 				cRenderer = new DevelopmentCardRenderer(built[i],sf::Transform(sprite_transforms[0]).translate(50.f+130.f*(i%14),350.f+160.f*(i/14)+50.f*(i/28)).scale(0.3f,0.3f));
 				this->devCardRenderers.push_back(cRenderer);	//Card Renderer
 			}
-			for (i=28;i<56 and i<28+tobuild.size();i++){
+			for (i=28;i<56 and i<28+(int) tobuild.size();i++){
 				cRenderer = new DevelopmentCardRenderer(tobuild[i],sf::Transform(sprite_transforms[0]).translate(50.f+130.f*(i%14),350.f+160.f*(i/14)+50.f*(i/28)).scale(0.3f,0.3f));
 				this->devCardRenderers.push_back(cRenderer);	//Card Renderer
 			}
@@ -290,23 +290,23 @@ namespace render {
 		
 		switch(this->affected_window){
 			case MAIN_WINDOW:
-				for (i=0;i<14 and i<tobuild.size();i++){
+				for (i=0;i<14 and i<(int) tobuild.size();i++){
 					cRenderer = new DevelopmentCardRenderer(tobuild[i],sf::Transform(sprite_transforms[0]).translate(350.f+70.f*(i%7),48.f+127.f*(i/7)).scale(0.2f,0.2f).scale(1.f,(431.f/375.f)));
 					this->devCardRenderers.push_back(cRenderer);	//Card Renderer
 				}
 				break;
 			case DRAFTING_WINDOW:
-				for (i=0;i<7 and i<drafted.size();i++){
+				for (i=0;i<7 and i<(int) drafted.size();i++){
 					cRenderer = new DevelopmentCardRenderer(drafted[i],sf::Transform(sprite_transforms[0]).translate(350.f+180.f*(i),10.f).scale(0.35f,0.35f).scale(1.f,(431.f/375.f)));
 					this->devCardRenderers.push_back(cRenderer);	//Card Renderer
 				}
 				break;
 			case PLAYER_INFO:
-				for (i=0;i<28 and i<built.size();i++){
+				for (i=0;i<28 and i<(int) built.size();i++){
 					cRenderer = new DevelopmentCardRenderer(built[i],sf::Transform(sprite_transforms[0]).translate(50.f+130.f*(i%14),350.f+160.f*(i/14)+50.f*(i/28)).scale(0.3f,0.3f));
 					this->devCardRenderers.push_back(cRenderer);	//Card Renderer
 				}
-				for (i=28;i<56 and i<28+tobuild.size();i++){
+				for (i=28;i<56 and i<28+(int) tobuild.size();i++){
 					cRenderer = new DevelopmentCardRenderer(tobuild[i],sf::Transform(sprite_transforms[0]).translate(50.f+130.f*(i%14),350.f+160.f*(i/14)+50.f*(i/28)).scale(0.3f,0.3f));
 					this->devCardRenderers.push_back(cRenderer);	//Card Renderer
 				}
